+++ app/src/test/java/com/codetutor/countryinfoapp/ExampleUnitTest.kt
@@ -1,11 +1,6 @@
 package com.codetutor.countryinfoapp
 
-import com.codetutor.countryinfoapp.database.Converters
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
-import org.junit.After
 import org.junit.Assert.assertEquals
-import org.junit.Before
 import org.junit.Test
 
 /**
